# docs/knowledge-graph/10-openapi.yaml
openapi: 3.1.0
info:
  title: TP Risk Management SMS — Safety Knowledge Graph Platform API
  version: 0.10.0-draft
  description: >
    DRAFT — controlled design document, requires approval before implementation.
    Backend REST surface for the Safety Knowledge Graph + AI Risk Intelligence Platform.
    Postgres is the system of record (03-postgresql-schema.sql); every write here is
    mirrored into Neo4j by the Graph Sync Service (01-enterprise-knowledge-graph-specification.md §4).
    See docs/PLATFORM_ARCHITECTURE_V2.md and docs/knowledge-graph/*.md for the design
    documents this contract implements.
    v0.2.0: Design Baseline v1.1 amendment — Emergency Planning (ACR-002) + Competency
    Management (ACR-003, supersedes ACR-001), approved 2026-08-04. Additive only —
    no removed/changed paths or schemas (02-development-standards.md §2 MINOR bump rule).
    v0.3.0: ACR-004 (approved 2026-08-11) — Investigation (1:1, INVESTIGATED_AS),
    incident-hazards linking (N:N, REVEALS, Option A bare-reference-list per ACR-004 §14),
    and incident-scoped Evidence. Drafted against ADR-003 (Investigation and Action are
    independent siblings of Incident, not a chain). Additive only — no removed/changed
    paths or schemas.
servers:
  - url: https://api.{environment}.vrtp-sms.example.net/v1
    variables:
      environment:
        default: dev
        enum: [dev, staging, prod]
security:
  - bearerAuth: []
...
paths:
...
  /risks/{riskId}/controls:
    parameters: [ { name: riskId, in: path, required: true, schema: { type: string, format: uuid } } ]
    get:
      tags: [Controls]
      summary: List controls mitigating this risk
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Control' } } } } }
    post:
      tags: [Controls]
      summary: Create a control candidate for this risk (pre 3-gate classification)
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/ControlInput' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Control' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /controls/{id}:
    parameters: [ { $ref: '#/components/parameters/id' } ]
    get:
      tags: [Controls]
      summary: Get a control
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Control' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /controls/{id}/gate-test:
    parameters: [ { $ref: '#/components/parameters/id' } ]
    post:
      tags: [Controls]
      summary: Submit 3-gate test answers (08-critical-control-assurance-model.md §2)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [gate_1, gate_2, gate_3]
              properties:
                gate_1: { type: boolean }
                gate_2: { type: boolean }
                gate_3: { type: boolean }
                is_verification_check: { type: boolean, description: "Follow-up if any gate failed" }
      responses:
        '200':
          description: Resulting classification
          content:
            application/json:
              schema:
                type: object
                properties:
                  classification: { type: string, enum: [Control, Support, Verification] }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /controls/{id}/eia-test:
    parameters: [ { $ref: '#/components/parameters/id' } ]
    post:
      tags: [Controls]
      summary: Record the Effective/Independent/Auditable quick assessment (08-critical-control-assurance-model.md §4a, Guide Table 2) — distinct from the 3-gate classification test
      requestBody:
        content: { application/json: { schema: { $ref: '#/components/schemas/ControlEiaInput' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Control' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /controls/{id}/critical-control-test:
    parameters: [ { $ref: '#/components/parameters/id' } ]
    post:
      tags: [Controls, CriticalControls]
      summary: Submit critical control test answer (08 §3) — only valid if classification = Control
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [is_critical]
              properties:
                is_critical: { type: boolean }
      responses:
        '201':
          description: CriticalControl created
          content: { application/json: { schema: { $ref: '#/components/schemas/CriticalControl' } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: "Control is not classified 'Control' — cannot be tested for criticality" }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /incidents/{id}/actions:
    parameters: [ { $ref: '#/components/parameters/id' } ]
    get:
      tags: [Incidents, Actions]
      summary: List Actions linked to this incident's linked-CAR roster (TRIGGERS, 1:N — safety.incident_actions; V1 fCARs)
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Action' } } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    post:
      tags: [Incidents, Actions]
      summary: >
        Link an existing Action to this incident's roster (ACR-006 Option A — bare reference,
        no IncidentAction schema; safety.incident_actions has no columns beyond the
        incident_id/action_id composite key. Does not create an Action — use POST /actions
        for creation.)
      requestBody: { content: { application/json: { schema: { type: object, required: [action_id], properties: { action_id: { type: string, format: uuid } } } } } }
      responses:
        '201': { description: Linked }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /incidents/{id}/actions/{actionId}:
    parameters:
      - { $ref: '#/components/parameters/id' }
      - { name: actionId, in: path, required: true, schema: { type: string, format: uuid } }
    delete:
      tags: [Incidents, Actions]
      summary: Unlink an Action from this incident's roster
      responses:
        '204': { description: Unlinked }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /actions:
    get:
      tags: [Actions]
      parameters: [ { $ref: '#/components/parameters/limit' }, { $ref: '#/components/parameters/offset' }, { in: query, name: status, schema: { type: string, enum: [Open, In Progress, Closed] } } ]
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Action' } } } } } }
    post:
      tags: [Actions]
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/ActionInput' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Action' } } } }
        '422': { $ref: '#/components/responses/ValidationError' }
  /actions/{id}:
    parameters: [ { $ref: '#/components/parameters/id' } ]
    patch:
      tags: [Actions]
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/ActionInput' } } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Action' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "AuthN/AuthZ provider TO BE CONFIRMED (architecture §2 AuthZ component) — Azure AD / Entra ID is the natural default given Azure hosting (architecture §8) but not yet decided."
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    limit:
      name: limit
      in: query
      schema: { type: integer, default: 50, maximum: 200 }
    offset:
      name: offset
      in: query
      schema: { type: integer, default: 0 }
  responses:
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    ValidationError:
      description: Request failed validation
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
    ConceptRef:
      type: object
      description: Reference to an ontology concept (03-postgresql-schema.sql ontology.concepts)
      properties:
        concept_id: { type: string, format: uuid, nullable: true }
        pref_label: { type: string, description: "Read-only, resolved server-side", nullable: true }
...
    Control:
      allOf:
        - $ref: '#/components/schemas/ControlInput'
        - type: object
          properties:
            id: { type: string, format: uuid }
            classification: { type: string, enum: [Control, Support, Verification], readOnly: true, nullable: true }
            is_critical: { type: boolean, readOnly: true }
    ControlInput:
      type: object
      required: [risk_id, description, control_type]
      properties:
        risk_id: { type: string, format: uuid }
        description: { type: string }
        control_type: { type: string, enum: [Prevention, Mitigation] }
        hierarchy: { $ref: '#/components/schemas/ConceptRef' }
        owner_person_id: { type: string, format: uuid }
        effectiveness_rating: { type: string, nullable: true }
    CriticalControl:
      type: object
      properties:
        control_id: { type: string, format: uuid }
        farsi_functionality: { type: integer, minimum: 1, maximum: 5 }
        farsi_availability: { type: integer, minimum: 1, maximum: 5 }
        farsi_reliability: { type: integer, minimum: 1, maximum: 5 }
        farsi_survivability: { type: integer, minimum: 1, maximum: 5 }
        farsi_interdependency: { type: integer, minimum: 1, maximum: 5, description: "Corrected from 'interaction' — confirmed against WHSQ Guide 2021 §5/§9.2.2.1 (08-critical-control-assurance-model.md §4b)" }
        farsi_score: { type: number, format: float, readOnly: true }
        eia_effective: { type: boolean, description: "Guide Table 2 — 08-critical-control-assurance-model.md §4a" }
        eia_independent: { type: boolean }
        eia_auditable: { type: boolean }
        health_state: { type: string, enum: [Verified, Healthy, Degraded, Overdue, Unverified], readOnly: true, description: "Computed — 08-critical-control-assurance-model.md §6" }
...
    ControlEiaInput:
      type: object
      description: "08-critical-control-assurance-model.md §4a — Guide Table 2"
      properties:
        eia_effective: { type: boolean }
        eia_independent: { type: boolean }
        eia_auditable: { type: boolean }
...
    Action:
      allOf:
        - $ref: '#/components/schemas/ActionInput'
        - type: object
          properties: { id: { type: string, format: uuid } }
    ActionInput:
      type: object
      required: [description]
      properties:
        source_type: { $ref: '#/components/schemas/ConceptRef' }
        source_id: { type: string, format: uuid }
        description: { type: string }
        root_cause_category: { $ref: '#/components/schemas/ConceptRef' }
        priority: { type: string, enum: [Critical, High, Medium, Low], nullable: true }
        assigned_to_person_id: { type: string, format: uuid }
        due_date: { type: string, format: date }
        status: { type: string, enum: [Open, In Progress, Closed] }
        effectiveness_review: { type: string, enum: ["Not Reviewed", Effective, "Partially Effective", Ineffective] }
        completion_date: { type: string, format: date }
        notes: { type: string, nullable: true }
...